Backend/requirements.txt
flask
werkzeug
pillow
numpy
tensorflow-cpu   # or `tensorflow` if you have GPU and supported OS
gunicorn
opencv-python